Split logic into modular chunks
